Move DVC Tracked tree into repository directory structure
